Removes interfaces and exposes functions directly instead Moves sync from cli to individual packages
