+++ .changeset/lemon-toys-camp.md
@@ -1,0 +1,7 @@
+---
+"ynab-sync-core": minor
+"ynab-sync-st-george-au": minor
+"ynab-sync-westpac-au": minor
+---
+
+Refactor to remove interfaces and expose functions instead
