+++ packages/core/package.json
@@ -20,6 +20,7 @@
     "lodash": "^4.17.21",
     "ofx": "^0.5.0",
     "string-template": "^1.0.0",
-    "ynab": "^1.25.0"
+    "ynab": "^1.25.0",
+    "os-locale": "^5.0.0"
   }
 }
