--- packages/core/src/exporter/ITransactionExporter.ts
@@ -1,7 +1,0 @@
-export interface ITransactionExporter<TInput, TOutput> {
-  export(input: TInput): Promise<TOutput>;
-}
-
-export type FileTransactionExportOutput = {
-  filePath: string;
-};
