--- packages/core/src/importer/ITransactionImporter.ts
@@ -1,15 +1,0 @@
-import { TransactionDetail } from "ynab";
-
-export type TransactionImportResults = {
-  transactionsCreated: TransactionDetail[];
-  transactionsUpdated: TransactionDetail[];
-  transactionsUnchanged: TransactionDetail[];
-};
-
-export interface ITransactionImporter {
-  import(
-    budgetId: string,
-    accountId: string,
-    transactions: TransactionDetail[]
-  ): Promise<TransactionImportResults>;
-}
