--- packages/core/src/importer/index.ts
@@ -1,2 +1,0 @@
-export * from './ITransactionImporter';
-export * from './YnabTransactionImporter';
