--- packages/core/src/parser/ITransactionParser.ts
@@ -1,5 +1,0 @@
-import { TransactionDetail } from 'ynab';
-
-export interface ITransactionParser {
-  parse(accountId: string, filePath: string): TransactionDetail[];
-}
