--- packages/core/src/parser/index.ts
@@ -1,3 +1,0 @@
-export * from "./ITransactionParser";
-export * from "./OfxTransactionParser";
-export * from "./CsvTransactionParser";
