+++ packages/st-george-au/src/export/index.ts
@@ -1,3 +1,2 @@
-export * from "./WestpacTransactionExporter";
 export * from "./login";
 export * from "./exportTransactions";
