--- packages/st-george-au/src/exporter/index.ts
@@ -1,1 +1,0 @@
-export * from "./StGeorgeTransactionExporter";
