+++ packages/st-george-au/src/sync/index.ts
@@ -1,0 +1,1 @@
+export * from "./syncTransactions";
