+++ packages/westpac-au/src/index.ts
@@ -1,1 +1,2 @@
-export * from "./exporter";
+export * from "./export";
+export * from "./sync";
